+++ RELEASES.md
@@ -1,3 +1,11 @@
+# Release 0.14.1 (2019-08-24) (Candidates)
+
+* More BLAS
+    * Vector & Vector Addition/Subtraction (`daxpy`)
+    * Vector & Scalar multiplication (`dscal`)
+    * Vector & Vector dot product (`ddot`)
+    * Vector Euclidean norm (`dnrm2`)
+
 # Release 0.14.0 (2019-08-24)
 
 * **[Important]** Finally, BLAS integrated.